fixed tests after refactor
