+++ __tests__/orders/list.js
@@ -1,8 +1,8 @@
 import {Orders} from "../../src/orders/orders";
 import {Order} from "../../src/orders/models";
-import {BloomableApi} from "../../src/bloomable/BloomableApi";
+import {BloomableApi} from "../../src/bloomable/api";
 
-jest.mock("bloomablelogistics-server/src/bloomable/BloomableWebsite");
+jest.mock("bloomablelogistics-server/src/bloomable/api");
 
 describe("orders list", () => {
     BloomableApi.getOrders.mockImplementation(() => Promise.resolve([new Order()]))
